Solving issue with Reactive

+++ src/main/java/com/pfc2/weather/pruebatecnica/model/WeatherHistoryResponse.java
@@ -13,5 +13,5 @@ public class WeatherHistoryResponse {
     private String weather;
     private Double tempMin;
     private Double tempMax;
-    private Double humidity;
+    private Integer humidity;
 }
